refactor(blog): extract Article SEO logic into useArticleSeo composable
The blog detail page setup script had grown to roughly 160 lines, most
of it SEO bookkeeping — image variants, thumbnail and word-count
computeds, the useSeoMeta payload, the Article JSON-LD graph,
breadcrumbs, and the OG image template invocation. None of that touched
the actual view code.

Move the lot into composables/useArticleSeo.ts, which takes the blog
and author refs from useBlogArticle and registers all the side
effects internally. The page now consumes only `title` from the
composable and keeps the head-link merging it already did via
useBlogArticle.

Result: the page setup is ~15 lines, the Article schema is reusable if
another route ever needs to render an article preview, and the JSON-LD
output is byte-identical (verified with the seo-check gate).
